+++ samples/android/camerakit-sample/camerakit-sample-full/src/main/java/com/snap/camerakit/sample/PreviewActivity.kt
@@ -119,7 +119,7 @@ class PreviewActivity : AppCompatActivity(), LifecycleOwner {
                         exportedMediaUri = if (exportedMediaUri != null) {
                             exportedMediaUri
                         } else {
-                            tryInsertToMediaStore(mediaFile, mediaMimeType)
+                            generateContentUri(mediaFile)
                         }?.also { uri ->
                             shareExternally(uri, mediaMimeType)
                         }